requirements.txt
# Web Framework
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5

# Task Queue
celery==5.3.4
redis==5.0.1

# Scheduling
APScheduler==3.10.4

# HTTP Requests
requests==2.31.0

# Database
SQLAlchemy==2.0.23

# Environment Variables
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0

# Development
flask-shell-ipython==1.4.0

# Production
gunicorn==21.2.0
supervisor==4.2.5

# Monitoring
flower==2.0.1
prometheus-client==0.19.0
psutil==5.9.6

# Logging
structlog==23.2.0

# Security and Additional Libraries
cryptography==41.0.7
